Consolidate config checking logic and use better exception.
