+++ community/kernel/src/main/java/org/neo4j/kernel/impl/spi/SimpleKernelContext.java
@@ -21,9 +21,7 @@ package org.neo4j.kernel.impl.spi;
 
 import java.io.File;
 
-import org.neo4j.graphdb.factory.GraphDatabaseSettings;
 import org.neo4j.io.fs.FileSystemAbstraction;
-import org.neo4j.kernel.configuration.Config;
 import org.neo4j.kernel.impl.factory.DatabaseInfo;
 import org.neo4j.kernel.impl.util.DependencySatisfier;
 
@@ -65,11 +63,5 @@ public class SimpleKernelContext implements KernelContext
     public DependencySatisfier dependencySatisfier()
     {
         return satisfier;
-    }
-
-    @Override
-    public boolean customIOConfigurationUsed( Config config )
-    {
-        return config.get( GraphDatabaseSettings.pagecache_swapper ) != null;
     }
 }
